Move packs js to entrypoint js
